refactor(lib/webidl2): transfer primitive_type() to helper.js
