+++ lib/productions/helpers.js
@@ -1,3 +1,5 @@
+import { Type } from "./type.js";
+
 /**
  * @param {string} identifier
  */
@@ -61,10 +63,42 @@ export function const_data({ type, value }) {
       return { type: "dictionary" };
     case "decimal":
     case "integer":
-        return { type: "number", value };
+      return { type: "number", value };
     case "string":
       return { type: "string", value: value.slice(1, -1) };
     default:
       return { type };
+  }
+}
+
+/**
+ * @param {import("../tokeniser").Tokeniser} tokeniser
+ */
+export function primitive_type(tokeniser) {
+  function integer_type() {
+    const prefix = tokeniser.consume("unsigned");
+    const base = tokeniser.consume("short", "long");
+    if (base) {
+      const postfix = tokeniser.consume("long");
+      return new Type({ source, tokens: { prefix, base, postfix } });
+    }
+    if (prefix) tokeniser.error("Failed to parse integer type");
+  }
+
+  function decimal_type() {
+    const prefix = tokeniser.consume("unrestricted");
+    const base = tokeniser.consume("float", "double");
+    if (base) {
+      return new Type({ source, tokens: { prefix, base } });
+    }
+    if (prefix) tokeniser.error("Failed to parse float type");
+  }
+
+  const { source } = tokeniser;
+  const num_type = integer_type(tokeniser) || decimal_type(tokeniser);
+  if (num_type) return num_type;
+  const base = tokeniser.consume("boolean", "byte", "octet");
+  if (base) {
+    return new Type({ source, tokens: { base } });
   }
 }
