+++ lib/webidl2.js
@@ -1,6 +1,6 @@
 "use strict";
 
-import { const_data, const_value, list, unescape } from "./productions/helpers.js";
+import { const_data, const_value, list, unescape, primitive_type } from "./productions/helpers.js";
 import { argumentNameKeywords, stringTypes, Tokeniser } from "./tokeniser.js";
 import { ArrayBase } from "./productions/array-base.js";
 import { Base } from "./productions/base.js";
@@ -37,34 +37,6 @@ function parseByTokens(tokeniser, options) {
 
   function unconsume(position) {
     return tokeniser.unconsume(position);
-  }
-
-  function integer_type() {
-    const prefix = consume("unsigned");
-    const base = consume("short", "long");
-    if (base) {
-      const postfix = consume("long");
-      return new Type({ source, tokens: { prefix, base, postfix } });
-    }
-    if (prefix) error("Failed to parse integer type");
-  }
-
-  function float_type() {
-    const prefix = consume("unrestricted");
-    const base = consume("float", "double");
-    if (base) {
-      return new Type({ source, tokens: { prefix, base } });
-    }
-    if (prefix) error("Failed to parse float type");
-  }
-
-  function primitive_type() {
-    const num_type = integer_type() || float_type();
-    if (num_type) return num_type;
-    const base = consume("boolean", "byte", "octet");
-    if (base) {
-      return new Type({ source, tokens: { base } });
-    }
   }
 
   function type_suffix(obj) {
@@ -118,7 +90,7 @@ function parseByTokens(tokeniser, options) {
   }
 
   function single_type(typeName) {
-    let ret = GenericType.parse(typeName) || primitive_type();
+    let ret = GenericType.parse(typeName) || primitive_type(tokeniser);
     if (!ret) {
       const base = consume(ID, ...stringTypes);
       if (!base) {
@@ -319,7 +291,7 @@ function parseByTokens(tokeniser, options) {
       if (!tokens.base) {
         return;
       }
-      let idlType = primitive_type();
+      let idlType = primitive_type(tokeniser);
       if (!idlType) {
         const base = consume(ID) || error("No type for const");
         idlType = new Type({ source, tokens: { base } });
